Split up proto::streams
